Create module index
